Update the timeseries module
This update inlcude addition of new variables:
- variable_known_mean
- variable_known_standard_deviation
- dump_trajectory
- dump_trajectory_fp
By addition of `variable_known_mean` & `variable_known_standard_deviation`
we provide extra control on the convergence package. It is beneficial
when we have theoretical value and want to check if the algorithm
converges within the valid range.

The code has been updated, and messages have been updated to match.
Based on the request now it is able to dump the trajectory.